backend/requirements.txt
# =============================================================================
# Archivo: requirements.txt
# Carpeta: backend/
# Propósito: Lista de dependencias de Python para el API Django.
# Cómo se usa: Docker (y entornos locales) ejecutan `pip install -r requirements.txt`.
# =============================================================================
Django>=5.1,<6
djangorestframework>=3.15,<4
django-cors-headers>=4.6,<5
psycopg2-binary>=2.9,<3
gunicorn>=23,<24
